feat: add welcome screen with login and registration functionality
- Implemented WelcomeScreen widget for user authentication.
- Added login functionality with error handling and loading state.
- Integrated quick test login options for demo users.
- Created modal bottom sheet for login input.
- Added navigation to FunRegistrationWizardScreen for user registration.

test: add unit tests for user profile serialization and preference catalog

- Added tests for UserProfile serialization and deserialization.
- Verified preference catalog contains dating intentions and diverse subcategories.
- Tested starter room generation based on user tastes.

feat: create script for generating starter furniture assets

- Developed a Python script to generate furniture images using PIL.
- Included various furniture items with isometric designs.
- Implemented saving of images with rotation variants for different placements.

+++ frontend/lib/features/lobby/components/isometric_furniture_preview.dart
@@ -6,15 +6,15 @@ import 'isometric_furniture_component.dart';
 
 class FurniturePreviewGame extends FlameGame {
   final FurnitureType type;
-  final int gw;
-  final int gh;
+  final double gw;
+  final double gh;
   final bool isDragged;
   final Sprite? sprite;
 
   FurniturePreviewGame({
     required this.type, 
-    this.gw = 1, 
-    this.gh = 1, 
+    this.gw = 1.0, 
+    this.gh = 1.0, 
     this.isDragged = false,
     this.sprite,
   });
@@ -26,13 +26,13 @@ class FurniturePreviewGame extends FlameGame {
     camera.viewfinder.anchor = Anchor.center;
     
     final furniture = IsometricFurnitureComponent(
-      gridX: 0,
-      gridY: 0,
+      gridX: 0.0,
+      gridY: 0.0,
       gridWidth: gw,
       gridHeight: gh,
       type: type,
       sprite: sprite,
-    )..updateGridPosition(0, 0);
+    )..updateGridPosition(0.0, 0.0);
 
     if (isDragged) {
       furniture.isSelected = true;
@@ -66,8 +66,8 @@ Widget previewBed() {
   return GameWidget(
     game: FurniturePreviewGame(
       type: FurnitureType.bed,
-      gw: 1,
-      gh: 2,
+      gw: 1.0,
+      gh: 2.0,
     ),
   );
 }
@@ -77,8 +77,8 @@ Widget previewBedRotated() {
   return GameWidget(
     game: FurniturePreviewGame(
       type: FurnitureType.bed,
-      gw: 2,
-      gh: 1,
+      gw: 2.0,
+      gh: 1.0,
     ),
   );
 }
@@ -88,8 +88,8 @@ Widget previewBedDragged() {
   return GameWidget(
     game: FurniturePreviewGame(
       type: FurnitureType.bed,
-      gw: 1,
-      gh: 2,
+      gw: 1.0,
+      gh: 2.0,
       isDragged: true,
     ),
   );
